+++ server/src/main/java/org/opensearch/index/knn/codec/package-info.java
@@ -7,4 +7,4 @@
  */
 
 /** Common interfaces for knn codecs to extend. */
-package org.opensearch.index.codec.knn;
+package org.opensearch.index.knn.codec;
